settings: Add button to reset channel notifications to default.
This commit adds a "Reset to default notifications" button in the
Notifications menu, to the right of the channel name, for which a
user has selected custom notifications and a confirmation
dialog to confirm this action.

Fixes #19860.
