+++ web/templates/confirm_dialog/confirm_reset_stream_notifications.hbs
@@ -1,0 +1,6 @@
+<p>
+    {{#tr}}
+    Are you sure you want to reset notifications for <z-stream></z-stream>?
+    {{#*inline "z-stream"}}{{> ../inline_decorated_channel_name stream=sub show_colored_icon=false}}{{/inline}}
+    {{/tr}}
+</p>
